extract new password component
